Move `LoadJsonFile` from `Features.Utils` to `Utils.JsonLoader` (khanshoaib3/stardew-access#112).
