+++ stardew-access/Patches/TitleMenuPatches/CharacterCustomizationMenuPatches.cs
@@ -1,7 +1,7 @@
-using System.Text.Json;
 using StardewValley;
 using StardewValley.Menus;
-using static stardew_access.Features.Utils;
+using System.Text.Json;
+using static stardew_access.Utils.JsonLoader;
 using static stardew_access.Utils.ColorMatcher;
 
 namespace stardew_access.Patches
